+++ changelog.md
@@ -1,2 +1,6 @@
+## 2024.2_v1.1.0
+* Updated to be compatible with Libero 2024.2
+* Removed product models: me-mp1-460-1si-d4e and me-mp1-250-ees-d3e
+
 ## 2022.2_v1.0.0
   * First Release